requirements.txt
requests==2.23.0
pytest==5.3.2
pytest-cov==2.8.1
pytest-metadata==1.8
pytest-html==2.0.1
pytest-clarity==0.2.0a1
pytest-xdist==1.31.0
lxml==4.4.2
allure-pytest==2.8.6
ipdb==0.12.3
pyyaml==5.3
flake8==3.7.9
imgcompare==2.0.1
diffimg==0.2.3
